chore: refactor some props
