+++ ios/Video/RCTVideoManager.swift
@@ -52,7 +52,26 @@ class RCTVideoManager: RCTViewManager {
     @objc public func replaceAdTagParameters(_ node: NSNumber, payload: NSDictionary) {
         DispatchQueue.main.async {
             let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
-            component?.replaceAdTagParameters(payload: payload)
+            
+            var adTagParameters = [String: Any]()
+            var startDate: Date?
+            var endDate: Date?
+
+            if let adTagParametersPayload = payload.value(forKey: "adTagParameters") as? [String: Any] {
+                adTagParameters = adTagParametersPayload
+            }
+
+            if let startDateInterval = payload.value(forKey: "startDate") as? Double {
+                startDate = Date(timeIntervalSince1970: startDateInterval)
+            }
+
+            if let endDateInterval = payload.value(forKey: "endDate") as? Double {
+                endDate = Date(timeIntervalSince1970: endDateInterval)
+            }
+            
+            component?.replaceAdTagParameters(adTagParameters: adTagParameters,
+                                              validFrom: startDate,
+                                              validUntil: endDate)
         }
     }
